feat: implement sudo mode

+++ svc-web-remix/app/routes/account/devices/$uid/delete.tsx
@@ -7,13 +7,18 @@ import { Card } from '~/components/retro-ui'
 import { getSessionData } from '~/utils/auth.server'
 import { sdk } from '~/utils/api.server'
 import type { GetUserDeviceByUidQuery } from '~/graphql/generated'
+import { loginRedirect } from '~/utils/redirect.server'
 
 export const action: ActionFunction = async ({ request, params }) => {
-  const { authToken } = await getSessionData(request)
-  invariant(authToken, 'Must be logged in')
+  const session = await getSessionData(request)
+
+  if (session.role === 'guest') {
+    return loginRedirect(request, session)
+  }
+
+  const { authToken } = session
 
   const { uid: userDeviceUID } = params
-  invariant(userDeviceUID, 'Must have params.uid')
 
   await sdk.deleteUserDevice(
     {
@@ -21,6 +26,7 @@ export const action: ActionFunction = async ({ request, params }) => {
     },
     {
       authorization: `Bearer ${authToken}`,
+      'x-hasura-role': 'user',
     },
   )
 
@@ -35,8 +41,13 @@ interface LoaderData {
 }
 
 export const loader: LoaderFunction = async ({ request, params }) => {
-  const { authToken } = await getSessionData(request)
-  invariant(authToken, 'Must be logged in')
+  const session = await getSessionData(request)
+
+  if (session.role === 'guest') {
+    return loginRedirect(request, session)
+  }
+
+  const { authToken } = session
 
   const { uid: userDeviceUID } = params
   invariant(userDeviceUID, 'Expected params.uid')
@@ -47,6 +58,7 @@ export const loader: LoaderFunction = async ({ request, params }) => {
     },
     {
       authorization: `Bearer ${authToken}`,
+      'x-hasura-role': 'user',
     },
   )
